+++ app/keys.php
@@ -1,12 +1,10 @@
 <?php
 
-use LaravelLang\Lang\Processors\Json;
-use LaravelLang\Lang\Processors\Php;
+use LaravelLang\Lang\Processors\Process;
 
 require __DIR__ . '/../vendor/autoload.php';
 
 /** @var \LaravelLang\Lang\Application $app */
 $app = require_once __DIR__ . '/bootstrap/app.php';
 
-$app->processor(Json::make());
-$app->processor(Php::make());
+$app->processor(Process::make());