+++ app/tests/PhpTest.php
@@ -3,14 +3,12 @@
 namespace Tests;
 
 use Helldar\PrettyArray\Services\File as Pretty;
-use Helldar\Support\Facades\Helpers\Filesystem\Directory;
-use Helldar\Support\Facades\Helpers\Filesystem\File;
 
 final class PhpTest extends TestCase
 {
     public function testPhp(): void
     {
-        foreach ($this->files() as $filename) {
+        foreach ($this->files('.php') as $filename) {
             $source = $this->source($filename);
 
             foreach ($this->locales() as $locale) {
@@ -27,18 +25,10 @@ final class PhpTest extends TestCase
         }
     }
 
-    protected function files(): array
-    {
-        return File::names($this->source_path, static fn ($filename) => str_ends_with($filename, '.php'));
-    }
-
-    protected function locales(): array
-    {
-        return Directory::names($this->target_path);
-    }
-
     protected function load(string $path): array
     {
-        return Pretty::make()->load($path);
+        $items = Pretty::make()->load($path);
+
+        return $this->correctValues($items);
     }
 }