+++ docs/changelog.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
  * feature: [nb] #1747 auth.php pagination.php passwords.php validation-inline.php validation.php json file
+ * feature: #1748 Split translation keys into Laravel packages
 
 ## [9.1.2] - 2021-05-14